feat(controller): implement RegistrationController and adjust services
Adjusted related services to ensure RegistrationController works correctly.

+++ src/main/java/com/eventManagement/eventManagement/repository/RegistrationRepository.java
@@ -16,6 +16,6 @@ public interface RegistrationRepository extends JpaRepository<Registration,Long>
 
     Page<Registration> findByEventId(Long eventId, Pageable pageable);
     Page<Registration> findByUserId(Long userId, Pageable pageable);
-    Long countByEventId(Long eventId);
+    int countByEventId(Long eventId);
     boolean existsByUserIdAndEventId(Long userId, Long eventId);
 }
